requirements.txt
scrapy
python-dotenv
psycopg2-binary
scrapeops-scrapy-proxy-sdk